+++ CHANGELOG.md
@@ -120,6 +120,7 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 - We moved the preferences option "Warn about duplicates on import" option from the tab "File" to the tab "Import and Export". [koppor#570](https://github.com/koppor/jabref/issues/570)
 - When JabRef encounters `% Encoding: UTF-8` header, it is kept during writing (and not removed). [#8964](https://github.com/JabRef/jabref/pull/8964)
 - We replace characters which cannot be decoded using the specified encoding by a (probably another) valid character. This happens if JabRef detects the wrong charset (e.g., UTF-8 instead of Windows 1252). One can use the [Integrity Check](https://docs.jabref.org/finding-sorting-and-cleaning-entries/checkintegrity) to find those characters.
+- We moved the "add group" button from the bottom of the groups UI panel to the horizontal row of icon buttons at the top [koppor#529](https://github.com/koppor/jabref/issues/529)
 
 ### Fixed